// packages/db/migrations/snapshots/30b5699020534fc908287fdb9130f474a0d70d74facbbcb0361a242d40116d86/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Comment: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly authorId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Follows: {
      readonly followerId: CodecTypes['pg/text@1']['output'];
      readonly followingId: CodecTypes['pg/text@1']['output'];
    };
    readonly Like: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly imageUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly authorId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly PostTag: {
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly tagId: CodecTypes['pg/text@1']['output'];
    };
    readonly Tag: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
      readonly profilePictureUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly bio: CodecTypes['pg/text@1']['output'] | null;
      readonly isGuestSandbox: CodecTypes['pg/bool@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Comment: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly authorId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Follows: {
      readonly followerId: CodecTypes['pg/text@1']['input'];
      readonly followingId: CodecTypes['pg/text@1']['input'];
    };
    readonly Like: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly imageUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly authorId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly PostTag: {
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly tagId: CodecTypes['pg/text@1']['input'];
    };
    readonly Tag: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
      readonly profilePictureUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly bio: CodecTypes['pg/text@1']['input'] | null;
      readonly isGuestSandbox: CodecTypes['pg/bool@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly comment: {
      readonly authorId: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
    };
    readonly follows: {
      readonly followerId: CodecTypes['pg/text@1']['output'];
      readonly followingId: CodecTypes['pg/text@1']['output'];
    };
    readonly like: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly post: {
      readonly authorId: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly imageUrl: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly postTag: {
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly tagId: CodecTypes['pg/text@1']['output'];
    };
    readonly tag: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly user: {
      readonly bio: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly isGuestSandbox: CodecTypes['pg/bool@1']['output'];
      readonly profilePictureUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly username: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly comment: {
      readonly authorId: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
    };
    readonly follows: {
      readonly followerId: CodecTypes['pg/text@1']['input'];
      readonly followingId: CodecTypes['pg/text@1']['input'];
    };
    readonly like: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly post: {
      readonly authorId: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly imageUrl: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly postTag: {
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly tagId: CodecTypes['pg/text@1']['input'];
    };
    readonly tag: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly user: {
      readonly bio: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly isGuestSandbox: CodecTypes['pg/bool@1']['input'];
      readonly profilePictureUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly username: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;